+++ src/main/java/org/folio/search/service/id/ResourceIdService.java
@@ -1,22 +1,18 @@
 package org.folio.search.service.id;
 
 import static java.lang.String.format;
-import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.folio.search.utils.LogUtils.collectionToLogMsg;
 import static org.opensearch.search.sort.SortBuilders.fieldSort;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.io.OutputStreamWriter;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.apache.commons.collections4.CollectionUtils;
 import org.folio.search.configuration.properties.StreamIdsProperties;
 import org.folio.search.cql.CqlSearchQueryConverter;
 import org.folio.search.exception.SearchServiceException;
@@ -26,7 +22,6 @@ import org.folio.search.model.types.StreamJobStatus;
 import org.folio.search.repository.ResourceIdsJobRepository;
 import org.folio.search.repository.ResourceIdsTemporaryRepository;
 import org.folio.search.repository.SearchRepository;
-import org.folio.spring.FolioExecutionContext;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -41,28 +36,16 @@ public class ResourceIdService {
   private final CqlSearchQueryConverter queryConverter;
   private final ResourceIdsJobRepository jobRepository;
   private final ResourceIdsTemporaryRepository idsTemporaryRepository;
-  private final FolioExecutionContext folioExecutionContext;
 
   /**
-   * Returns resource ids for passed cql query in text type.
+   * Streams resource IDs as JSON from the database for a given job.
+   * The job must be completed and have IDs prepared in a temporary table.
    *
-   * @param request      resource ids request as {@link CqlResourceIdsRequest} object
-   * @param outputStream output stream where text will be written in.
-   */
-  public void streamResourceIdsAsText(CqlResourceIdsRequest request, OutputStream outputStream) {
-    var writer = createOutputStreamWriter(outputStream);
-    streamResourceIds(request, ids -> writeRecordIdsToOutputStream(ids, writer));
-  }
-
-  /**
-   * Returns resource ids for passed cql query in json format.
-   * Should have a prepared job with ids in the database.
-   *
-   * @param jobId        async jobs id with prepared query
-   * @param outputStream output stream where json will be written in.
+   * @param jobId the ID of the async job with the prepared query
+   * @param outputStream the output stream to write the JSON data to
    */
   @Transactional
-  public void streamIdsFromDatabaseAsJson(String jobId, OutputStream outputStream) {
+  public void streamResourceIdsAsJson(String jobId, OutputStream outputStream) {
     log.debug("streamIdsFromDatabaseAsJson:: by [jobId: {}]", jobId);
 
     var job = jobRepository.getReferenceById(jobId);
@@ -95,7 +78,7 @@ public class ResourceIdService {
    * @param tenantId tenant id as {@link String} object
    */
   @Transactional
-  public void streamResourceIdsForJob(ResourceIdsJobEntity job, String tenantId) {
+  public void processResourceIdsJob(ResourceIdsJobEntity job, String tenantId) {
     log.debug("streamResourceIdsForJob:: by [job: {}, tenantId: {}]", job, tenantId);
     var tableName = job.getTemporaryTableName();
     try {
@@ -106,7 +89,7 @@ public class ResourceIdService {
 
       log.info("streamResourceIdsForJob:: Attempting to create table for ids [tableName: {}]", tableName);
       idsTemporaryRepository.createTableForIds(tableName);
-      streamResourceIds(request, idsList -> idsTemporaryRepository.insertIds(idsList, tableName));
+      streamIdsFromSearch(request, idsList -> idsTemporaryRepository.insertIds(idsList, tableName));
       job.setStatus(StreamJobStatus.COMPLETED);
     } catch (Exception e) {
       log.warn("Failed to process resource ids job with id = {}, msg: {}", job.getId(), e.getMessage());
@@ -118,25 +101,7 @@ public class ResourceIdService {
     }
   }
 
-  /**
-   * Returns resource ids for passed cql query in json type.
-   *
-   * @param request      resource ids request as {@link CqlResourceIdsRequest} object
-   * @param outputStream output stream where json will be written in.
-   */
-  public void streamResourceIdsAsJson(CqlResourceIdsRequest request, OutputStream outputStream) {
-    processStreamToJson(outputStream, (json, counter) ->
-      streamResourceIds(request, ids -> {
-        counter.addAndGet(ids.size());
-        writeRecordIdsToOutputStream(ids, json);
-      }));
-  }
-
-  protected OutputStreamWriter createOutputStreamWriter(OutputStream outputStream) {
-    return new OutputStreamWriter(outputStream, UTF_8);
-  }
-
-  private void streamResourceIds(CqlResourceIdsRequest request, Consumer<List<String>> idsConsumer) {
+  private void streamIdsFromSearch(CqlResourceIdsRequest request, Consumer<List<String>> idsConsumer) {
     log.info("streamResourceIds:: by [query: {}, resource: {}]", request.getQuery(), request.getResource());
 
     var searchSource = queryConverter
@@ -166,44 +131,6 @@ public class ResourceIdService {
     } catch (IOException e) {
       throw new SearchServiceException(
         format("Failed to write data into json [reason: %s]", e.getMessage()), e);
-    }
-  }
-
-  private static void writeRecordIdsToOutputStream(List<String> recordIds, JsonGenerator json) {
-    log.debug("writeRecordIdsToOutputStream:: by [recordIds: {}, json]", collectionToLogMsg(recordIds));
-
-    if (CollectionUtils.isEmpty(recordIds)) {
-      return;
-    }
-
-    try {
-      for (var recordId : recordIds) {
-        json.writeStartObject();
-        json.writeStringField("id", recordId);
-        json.writeEndObject();
-      }
-      json.flush();
-    } catch (IOException e) {
-      throw new SearchServiceException(
-        format("Failed to write to id value into json stream [reason: %s]", e.getMessage()), e);
-    }
-  }
-
-  private static void writeRecordIdsToOutputStream(List<String> recordIds, OutputStreamWriter outputStreamWriter) {
-    log.debug("writeRecordIdsToOutputStream:: by [recordIds: {}, outputStreamWriter]", collectionToLogMsg(recordIds));
-
-    if (CollectionUtils.isEmpty(recordIds)) {
-      return;
-    }
-
-    try {
-      for (var recordId : recordIds) {
-        outputStreamWriter.write(recordId + '\n');
-      }
-      outputStreamWriter.flush();
-    } catch (IOException e) {
-      throw new SearchServiceException(
-        format("Failed to write id value into output stream [reason: %s]", e.getMessage()), e);
     }
   }
 }